Chore: Merge two layouts into one layout, change album UI, remove the previous index design & make search page into the new index page, added profile page
