+++ packages/frontend/src/components/dynamic/band/BandBreadcrumb.tsx
@@ -1,6 +1,4 @@
-import Link from "next/link";
 import React from "react";
-import { MdKeyboardArrowRight } from "react-icons/md";
 
 interface BandBreadcrumb {
   original_band: string | undefined;
@@ -8,17 +6,8 @@ interface BandBreadcrumb {
 
 const BandBreadcrumb: React.FC<BandBreadcrumb> = ({ original_band }) => {
   return (
-    <div className="flex h-[35px] items-end gap-2 rounded-lg border-2 px-3 py-1 sm:flex">
-      <Link
-        href={"/search"}
-        className="text-neutral-500 decoration-1 hover:underline"
-      >
-        Home
-      </Link>
-      <div className="hidden h-[90%] items-center justify-center sm:flex">
-        <MdKeyboardArrowRight />
-      </div>
-      <span className="hidden sm:block">{original_band}</span>
+    <div className="flex h-[35px] items-center gap-2 rounded-lg border-2 px-3 py-1 text-xs sm:text-sm">
+      {original_band}
     </div>
   );
 };
