+++ packages/frontend/src/components/home/SearchAlbum.tsx
