+++ packages/frontend/src/components/home/SearchAlbumList.tsx
