+++ packages/frontend/src/components/home/SearchBand.tsx
