+++ packages/frontend/src/components/home/SearchBar.tsx
