+++ packages/frontend/src/components/home/SearchLoading.tsx
