+++ packages/frontend/src/components/layout/Header.tsx
@@ -3,33 +3,16 @@ import Link from "next/link";
 import React from "react";
 import { IoPerson } from "react-icons/io5";
 
-interface headerDataProps {
-  link: string;
-  icon: React.ReactNode;
-}
-
-const iconStyle = "h-[20px] w-[20px]";
-
-const headerData: headerDataProps[] = [
-  { link: "", icon: <IoPerson className={iconStyle} /> },
-];
-
 const Header = () => {
   return (
-    <nav className="fixed top-0 z-20 flex h-[70px] w-full items-center justify-between border bg-white p-[10px] px-5">
+    <nav className="fixed top-0 z-20 flex h-[50px] w-full items-center justify-between border bg-white p-[10px] px-5 sm:h-[70px]">
       <Link href={"/"} className="flex items-center justify-center">
-        <Image src={"/img/logo.png"} alt="logo" width={50} height={50} />
-        <h1 className="text truncate ps-5 text-2xl font-semibold">
-          FGA Worship
-        </h1>
+        <Image src={"/img/logo.png"} alt="logo" width={30} height={30} />
+        <h1 className="text truncate ps-5 font-semibold">FGA Worship</h1>
       </Link>
-      <div className="flex items-center gap-3">
-        {headerData.map((item, i) => (
-          <Link key={i} href={item.link} className="rounded border p-1">
-            {item.icon}
-          </Link>
-        ))}
-      </div>
+      <Link href="/profile" className="hidden p-1 sm:block">
+        <IoPerson />
+      </Link>
     </nav>
   );
 };
