+++ packages/frontend/src/components/layout/Layout.tsx
@@ -1,6 +1,8 @@
 import Head from "next/head";
 import React, { type ReactNode } from "react";
 import Header from "./Header";
+import Footer from "./Footer";
+import { UserProvider } from "@auth0/nextjs-auth0/client";
 // import Navigation from "./Navigation";
 
 interface LayoutProps {
@@ -9,7 +11,7 @@ interface LayoutProps {
 
 const Layout: React.FC<LayoutProps> = ({ children }) => {
   return (
-    <>
+    <UserProvider>
       <Head>
         <title>Song Bank</title>
         <meta name="" content="" />
@@ -17,8 +19,11 @@ const Layout: React.FC<LayoutProps> = ({ children }) => {
       </Head>
       <Header />
       {/* <Navigation /> */}
-      <main className="relative top-[70px] z-0 min-h-[91vh]">{children}</main>
-    </>
+      <main className="relative top-[50px] z-0 min-h-[91vh] sm:top-[70px]">
+        {children}
+      </main>
+      <Footer />
+    </UserProvider>
   );
 };
