+++ packages/frontend/src/components/profile/ProfileFavouriteSongs.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ProfileFavouriteSongs = () => {
+  return <div>ProfileFavouriteSongs</div>;
+};
+
+export default ProfileFavouriteSongs;
