--- packages/frontend/src/pages/login.tsx
@@ -1,8 +1,0 @@
-import Link from "next/link";
-import React from "react";
-
-const login = () => {
-  return <Link href="/api/auth/login">Login</Link>;
-};
-
-export default login;
